+++ src/main/java/lotto/controller/validator/ErrorMessage.java
@@ -7,7 +7,8 @@ enum ErrorMessage {
     NOT_DIVIDED_BY_LOTTO_TICKET_PRICE("로또 티켓 금액의 배수가 아닙니다."),
     NOT_PROPER_SIZE_OF_LOTTO("번호의 개수가 6개가 아닙니다."),
     NUMBER_NOT_IN_RANGE("올바른 범위의 수가 아닙니다."),
-    HAS_DUPLICATED_NUMBER("중복 값이 존재합니다.");
+    HAS_DUPLICATED_NUMBER("중복 값이 존재합니다."),
+    NOT_PROPER_OBJECT_TYPE("올바른 객체 타입이 아닙니다.");
 
     private static final String MESSAGE_HEADER = "[ERROR]";
     private static final String WHITE_SPACE = " ";
